# api/gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    An example OpenAPI specification for the Little PM application, referencing path parameters from components.
  title: Little PM API
  version: 1.0.0
servers:
- description: Production server
  url: https://app.little.pm
...
paths:
...
  /reports/{userUUID}:
    get:
      description: Retrieve a list of generated reports for the specified user.
      operationId: getReports
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Report'
                type: array
          description: Reports found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: User not found
      summary: List all reports for a user
      tags:
      - Report Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Report Service
    post:
      description: "Create a new report (e.g., pitch deck, marketing, financial)."
      operationId: createReport
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
          description: Report generated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: User not found
      summary: Generate a new report
      tags:
      - Report Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Report Service
  /reports/{userUUID}/{reportUUID}:
    get:
      description: Retrieve a specific report by UUID for a user.
      operationId: getReport
      parameters:
      - description: ID of the user
        explode: false
        in: path
        name: userUUID
        required: true
        schema:
          $ref: '#/components/schemas/userUUID'
        style: simple
      - explode: false
        in: path
        name: reportUUID
        required: true
        schema:
          example: 665c599d-5c8d-4d20-aaab-7ffaba150606
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
          description: Report found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Report or user not found
      summary: Get a specific report
      tags:
      - Report Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Report Service
...
components:
...
  schemas:
    userUUID:
      description: ID of the user
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
...
    Error:
      description: Error object
      example:
        message: Error message
      properties:
        message:
          description: Error message
          example: Error message
          type: string
      type: object
...
    ReportRequest:
      example:
        reportType: pitchDeck
        inputs: "{}"
      properties:
        reportType:
          enum:
          - pitchDeck
          - marketing
          - financial
          type: string
        inputs:
          description: Key-value inputs for the template
          type: object
      type: object
    Report:
      example:
        reportType: reportType
        reportUUID: 665c599d-5c8d-4d20-aaab-7ffaba150606
        url: url
      properties:
        reportUUID:
          description: ID of the report
          example: 665c599d-5c8d-4d20-aaab-7ffaba150606
          format: uuid
          type: string
        reportType:
          type: string
        url:
          description: URL to download the generated file
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http